+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.2.12] - Unreleased
+### Added
+- `options` option to `list` field.
+- New `init` option to all fields.
+
+### Fixed
+- Updated `Hono` to the latest version.
+
 ## [0.2.11] - 2024-02-27
 ### Added
 - Types for `Field.toJSON`.